Refactor fragments to mvvm
