+++ app/src/main/res/layout/fragment_building_list.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:clickable="true"
     android:focusable="true"
-    tools:context=".fragments.BuildingListFragment">
+    tools:context=".ui.buildingList.BuildingListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
