+++ app/src/main/res/layout/fragment_maps.xml
@@ -5,4 +5,4 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.MapsFragment" />
+    tools:context=".ui.maps.MapsFragment" />
