+++ app/src/main/res/layout/fragment_stock_list.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:clickable="true"
     android:focusable="true"
-    tools:context=".fragments.StockListFragment">
+    tools:context=".ui.stockList.StockListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/sRecyclerView"
